move cart.py to main,py, menu almost finish
