+++ cart.txt
@@ -1,4 +1,11 @@
-TEH TARIK, 5.0, 3, 
+TEH TARIK, 5.0, 2, n
 COFFEE, 0.0, 1, 
 BANANA BAR, 3.8, 1, 
 ESPRESSO, 5.5, 1, 
+FIRE CRACKER SAUSAGE, 7.5, 1, 
+CHOCOLATE, 7.0, 1, 
+CHOCOLATE MUFFIN, 6.5, 1, 
+NON COFFEE, 0.0, 1, 
+DOUBLE CHOCO CHIP COOKIES, 4.5, 1, 
+SALTED CARAMEL LATTE, 8.5, 1, 
+FRENCH VANILLA LATTE, 8.5, 1,
