+++ templates/home.html
@@ -16,7 +16,7 @@
         </form> -->
         
         
-        <h2>Hello, <br><i>username</i></h2>
+        <h2>Hello, <br><i>{{ username }}</i></h2>
         
     </header>
